Restructure classes
The old class 'Generator' is now devided into two classes - 'Oscillator' and 'Generator'.
The new 'Generator' represents a generator that sends a constant signal, 'Oscillator' sends a sine wave of fixed frequency - 440Hz.
'Oscillator' is a subclass of 'Generator' and all subclasses of the old 'Generator' are now subclasses of 'Oscillator'.
The 'Oscillator' class is in the file 'oscillators.py'.

The method 'wave_y' in all classes posessing it is now renamed to 'output', and the method 'mod_y' in 'FMOperator' is renamed to 'mod_out'

the method 'Generator.draw' takes time as an argument instead of cycles, the 'draw' method of other classes can also take time as an argument but they do not by default.