Restructure FM Operator and extend its functionality
FMOperator got devided into two classes: 'FMOperator' and 'LinearFMGenerator'.
'LinearFMGenerator' represents a sound generator whose freqency can be linearly modulated,

'FMOperator' represents a module that includes a 'LinearFMGenerator' instance as its sound generator, a mixer that sends a modulating signal to the generator, an amplifier and an envelope generator. The amplifier amplifies the generator and is modulated by the envelope generator. The operators output is the amplifiers output.

If specified 'FMOperator's sound generator can be an instance of a new class 'DXGenerator' that represents a generator that is modulated differently than 'LinearFMGenerator', but this type of modulation is commonly referred to as FM.
This type of modulation is present in the Yamaha DX synthesizer series

Bug fixes:

Corrected the 'Mixer's attributes such that each 'Mixer' instance has its own lists of inputs and levels instead of global lists.

'Gate' can now be initiated with an empty list.

'line' function in 'math_func.py' can now recieve points lying on a vertical line.

Added missing comments in 'generators.py' and 'oscillators.py'